Obj loading fixes
Updated obj loading so that faces with more than 3 vertices doesn't cause the program to crash. The problem of faces with no texture coordinate hasn't been fixed, however an obj file shouldn't have texture coordinates without listing the indices with the faces and the program checks if the file has texture coordinates.