Transparent objects+ scene object data structure decision
Updated render calls so that non-transparent objects are rendered first. Then transparent objects are sorted farthest to nearest to the camera, then rendered. Also decided to use vectors instead of linked lists for sceneobjects as we need to iterate through sceneobjects many times, but will be unlikely to need to move objects around within the vector