Screen space + whole string update
Updated the UI shader so that it uses the projection matrix passed in. Also created a orthographic matrix to pass to the shader that has space ranging from 0 to 1 in the x and y axis, instead of the default -1 to 1. Lastly, updated OutputString so that it actually outputs the given string