Camera collision
Created a camera class which handles moving the camera and checking if the camera will colliding with a rectCollider if it moves. Also moved various bits of code into the camera class (e.g. mouse movement). Made an overload for the SceneObject::GetAllObjects in the scenario that you don't need to know if they are transparent or not