Transparency bug fix
Fixed bug where objects weren't drawn in the correct order due to the camera pointer passed on creation not being initialised. Also added 2 alternative glass textures to make it clearer that the current solution works. Also changed the sorting of transparent objects to use a function so that we don't have to reverse the list afterwards