Lighting, cleanup and frame
Added models for a frame and a flat plane. Updated FishShader to have the same lighting code used in VertLighting. Updated the max number of lights to 10. Added a light dimming feature where all scene lights can be dimmed. Removed unused code in update. Redid lighting. Updated all objects to use a shader with lighting. Also added a painting in the right side room which uses a blank white texture and a brown material.