Being in rooms changes the string at the top of the screen
Updated RectCollider so that a collider can be a trigger collider. Also updated SceneObject so that it has a function to call if the camera is colliding with it and its collider is a trigger.
Updated the camera so that it only does collision checks with non-trigger objects. Also added a function that calls the OnTrigger function if it is colliding with a given object.
Added a Room class which sets the room string to its collision string when triggered. Also updated HelloGL so that the rooms are of type room and they have the correct colliders. Also added roomString so that the rooms can update the string