Cleaned up code
Moved Camera's GetViewMatrix to the header file.
Extracted part of mesh loading to a function and made sure to delete vertex normals.
Updated GLUTCallbacks so that it checks that helloGL is valid and so that the constant is replaced.
Removed todo list from HelloGL.
Deleted all objects in HelloGL.
Extracted setting uniforms to a function in HelloGL.
Also extracted setting up the map to a function.
Added a destructor to Skybox
Replaced a number with a constant in Penguin
Added a destructor in Font. Also cleaned up the UpdateVertexBuffer call.
Updated Texture2D so that if Load is called after a texture has already been loaded in, then the previous texture is deleted.
Updated Vao so that a SetUpAttributes is private